requirements.txt
PyGithub==1.55
tqdm==4.62.3
brotli
selenium